Map with responsive layout. Geolocation logic
