+++ src/components/common/Map/map.scss
@@ -1,0 +1,7 @@
+.current-location-marker {
+  position: fixed;
+}
+
+.current-location-icon {
+  transform: translate(14%, -30%);
+}
