+++ src/constants/constants.js
@@ -2,3 +2,5 @@
 
 export const SUPPORTED_LANGUAGES = ['en', 'es'];
 export const DEFAULT_LANGUAGE = 'en';
+
+export const DEFAULT_COORDINATES = { lat: -26.851311, lng: -65.702984 };
