+++ src/styles/styles.scss
@@ -140,17 +140,3 @@ button {
 .sidebar-content {
   width: 30%;
 }
-
-.map-location-marker {
-  color: $black;
-  background-color: $primary-blue;
-  padding: 1.2rem 1.8rem;
-  display: inline-flex;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-  border-radius: 100%;
-  transform: translate(-50%, -50%);
-  font-size: 1.2rem;
-  font-weight: 600;
-}
